+++ src/test/node/EntryPoint/minimal-and-default/fp-admin.yaml
@@ -1,4 +1,4 @@
-site_name: FuckPress
+site_name: ed9fe511-665c-48e9-ae2e-dec37ed18693
 description: A ultra minimalistic headless cms
 home_banner_image_url: >-
   https://github.com/podcastjs/podcastjs/assets/3322836/0d8bdbf1-78f0-4728-bcf4-6a6a25487dfe
