+++ webpage/index.html
@@ -226,7 +226,7 @@ https://www.iconfinder.com/korawan_m
   <script src="js/bootstrap.bundle.min.js"></script>
   <!-- script js -->
   <script src="js/app.js"></script>
-  <a class="github-fork-ribbon" href="https://github.com/fuckpress/fuckpress-core"  target="_blank" data-ribbon="Fork me on GitHub" title="Fork me on GitHub">Fork me on GitHub</a>
+  <a class="github-fork-ribbon" href="https://github.com/fuckpress/fuckpress-demo"  target="_blank" data-ribbon="Fork me on GitHub" title="Fork me on GitHub">Fork me on GitHub</a>
 </body>
 
 </html>
